.claude/skills/add-example/pyproject.toml
[project]
name = "add-example-thumbnails"
version = "1.0.0"
description = "Thumbnail generation utilities for add-example skill"
requires-python = ">=3.11"
dependencies = [
    "pypdfium2>=5.0.0",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
